+++ src/AckModule.ts
@@ -2,9 +2,6 @@ import {
   ModuleWithProviders, NgModule
 } from "@angular/core"
 
-//deprecate importing this
-import { HttpClientModule } from "@angular/common/http"
-
 import { CommonModule } from "@angular/common"
 //import { FormsModule }   from "@angular/forms"
 
@@ -16,20 +13,11 @@ const declarations = [...components, ...pipes]
 
 @NgModule({
   imports:[
-    CommonModule,
-    //FormsModule,
-
-    //deprecate this dependency by seperating AckApi
-    HttpClientModule
+    CommonModule
   ],
   declarations: declarations,
   //providers:providers,
   exports:[
-    //FormsModule,
-    
-    //deprecate this
-    HttpClientModule,
-    
     ...declarations,
    ]
 }) export class AckModule {