+++ src/index.ts
@@ -1,13 +1,13 @@
 export { AckModule } from "./AckModule"
-export { AckRouterModule } from "./AckRouterModule"
-export { RouteWatchReporter } from "./RouteWatchReporter"
+export { AckRouterModule } from "./modules/router/AckRouterModule"
+export { RouteWatchReporter } from "./modules/router/RouteWatchReporter"
 
 export { declarations as components } from "./declarations"
 export { declarations as pipes } from "./pipes"
 
 export { httpOptions } from "./httpOptions"
 export { ack } from "./ack"
-export { AckApi } from "./providers/AckApi"
+export { AckApi } from "./modules/offline/AckApi"
 export { AckArray } from "./directives/AckArray.directive"
 
 export * from "./providers"