+++ src/providers.ts
@@ -10,20 +10,8 @@ import { AckApp } from "./providers/AckApp.provider";
 export { Prompts } from "./providers/Prompts";
 import { Prompts } from "./providers/Prompts";
 
-export { AckOffline } from "./providers/AckOffline";
-import { AckOffline } from "./providers/AckOffline";
-
-export { AckCache } from "./providers/AckCache";
-import { AckCache } from "./providers/AckCache";
-
-export { AckQue } from "./providers/AckQue";
-import { AckQue } from "./providers/AckQue";
-
 export { ErrorLog } from "./providers/ErrorLog";
 import { ErrorLog } from "./providers/ErrorLog";
-
-//export * from "./providers/AckApi";
-import { AckApi } from "./providers/AckApi";
 
 export { UrlVars } from "./providers/UrlVars";
 import { UrlVars } from "./providers/UrlVars";
@@ -38,10 +26,6 @@ export const providers = [
   Log,
   ErrorLog,
   AckApp,
-  AckOffline,
-  AckCache,
-  AckQue,
-  AckApi,
   UrlVars,
   WindowService,
   DocumentService,